backend/requirements.txt
fastapi
uvicorn
strawberry-graphql
python-dotenv
openai
pandas
numpy
aiohttp
rank-bm25
spacy
scikit-learn
matplotlib
